refactor(federated): split RTI module tests
